add service implementation

+++ src/main/java/tomaszslowik1/beusableproject/repository/RoomRepositoryJsonMock.java
@@ -4,18 +4,19 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 
 import java.io.File;
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.util.*;
 
 public class RoomRepositoryJsonMock implements RoomRepository{
     @Override
-    public List<Integer> getGuestsList()  {
+    public List<BigDecimal> getGuestsList()  {
         File file = new File(
                 Objects.requireNonNull(this.getClass().getClassLoader().getResource("static/guests.json")).getFile()
         );
         ObjectMapper mapper = new ObjectMapper();
-        Integer[] guestsArray;
+        BigDecimal[] guestsArray;
         try {
-            guestsArray = mapper.readValue(file, Integer[].class);
+            guestsArray = mapper.readValue(file, BigDecimal[].class);
         } catch (IOException e) {
             return Collections.emptyList();
         }
